Separated template for changing password during activation
