+++ DependencyInjection/Configuration.php
@@ -130,6 +130,12 @@ class Configuration implements ConfigurationInterface
                     ->children()
                         ->scalarNode('login')->defaultValue('FSiAdminSecurityBundle:Security:login.html.twig')->end()
                         ->scalarNode('change_password')->defaultValue('FSiAdminSecurityBundle:Admin:change_password.html.twig')->end()
+                        ->arrayNode('activation')
+                            ->addDefaultsIfNotSet()
+                            ->children()
+                                ->scalarNode('change_password')->defaultValue('FSiAdminSecurityBundle:Activation:change_password.html.twig')->end()
+                            ->end()
+                        ->end()
                         ->arrayNode('password_reset')
                             ->addDefaultsIfNotSet()
                             ->children()
