+++ Resources/translations/FSiAdminSecurity.en.yml
@@ -41,9 +41,13 @@ admin:
       message:
         success: Your password has been successfully changed
   activation:
+    change_password_form:
+      button:
+        change_password: Activate account
     mail:
       subject: User Activation
       info: "To activate your account - please visit:"
     message:
+      change_password: Please setup a new password to activate your account
       change_password_success: Your password has been successfully changed and your account has been activated
       success: Your account has been successfully activated
