+++ Resources/translations/FSiAdminSecurity.pl.yml
@@ -41,9 +41,13 @@ admin:
       message:
         success: Twoje hasło zostało poprawnie zmienione
   activation:
+    change_password_form:
+      button:
+        change_password: Aktywuj konto
     mail:
       subject: Aktywacja użytkownika
       info: "Aby aktywować twoje konto proszę otworzyć link:"
     message:
+      change_password: Proszę ustawić nowe hasło aby aktywować konto
       change_password_success: Twoje hasło zostało poprawnie zmienione oraz twoje konto zostało aktywowane
       success: Twoje konto zostało poprawnie aktywowane
