+++ Resources/views/Activation/change_password.html.twig
@@ -1,0 +1,14 @@
+{% extends '@FSiAdminSecurity/layout.html.twig' %}
+
+{% block themes %}
+    {% form_theme form 'bootstrap_3_layout.html.twig' %}
+{% endblock %}
+
+{% block body %}
+    {{ parent() }}
+
+    {{ form_start(form, {'attr': {'class': 'form-signin'}}) }}
+    {{ form_widget(form) }}
+    <input type="submit" class="btn btn-lg btn-primary btn-block" value="{{ 'admin.activation.change_password_form.button.change_password'|trans({}, 'FSiAdminSecurity') }}" />
+    {{ form_end(form) }}
+{% endblock %}
