+++ Resources/views/layout.html.twig
@@ -45,16 +45,18 @@
         .form-signin .form-control:focus {
             z-index: 2;
         }
-        .form-signin input[type="text"] {
+        .form-signin input[name="_username"] {
             margin-bottom: -1px;
             border-bottom-left-radius: 0;
             border-bottom-right-radius: 0;
         }
-        .form-signin input[type="password"],
-        .form-signin input[type="email"] {
-            margin-bottom: 10px;
+        .form-signin input[name="_password"] {
+            margin-bottom: 15px;
             border-top-left-radius: 0;
             border-top-right-radius: 0;
+        }
+        .flash {
+            margin: 0 15px;
         }
     </style>
     <!-- HTML5 shim and Respond.js IE8 support of HTML5 elements and media queries -->
@@ -71,7 +73,7 @@
         {% if app.request.hasPreviousSession %}
             {% for type, messages in app.session.flashbag.all() %}
                 {% for message in messages %}
-                    <div class="alert flash {{ type }}">
+                    <div class="alert flash alert-{{ type }}">
                         {{ message|trans({}, 'FSiAdminSecurity') }}
                     </div>
                 {% endfor %}
