+++ features/fixtures/project/app/Resources/views/activation.html.twig
@@ -1,0 +1,5 @@
+{% extends 'FSiAdminSecurityBundle:Activation:change_password.html.twig' %}
+
+{% block themes %}
+    {% form_theme form '::form_theme.html.twig' %}
+{% endblock %}
